requirements.txt
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
db-dtypes>=1.1.1
pandas>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
